correction

+++ exec/src/free/free_follow.c
@@ -38,11 +38,11 @@
 		free(glob->cmd);
 }*/
 
-void	free_tab(char **tab, int lenght)
+void	free_tab(char **tab, size_t lenght)
 {
 	if (!tab)
 		return ;
-	printf("info lenght = %d\n", lenght);
+	printf("info lenght = %zu\n", lenght);
 	if (lenght == 0)
 	{
 		free(tab[0]);
